Refactor s3 connectors 	* Make separate s3 module in impl.
