+++ src/connectors.rs
@@ -1157,8 +1157,8 @@ pub fn builtin_connector_types() -> Vec<Box<dyn ConnectorBuilder + 'static>> {
         Box::new(impls::ws::server::Builder::default()),
         Box::new(impls::elastic::Builder::default()),
         Box::new(impls::crononome::Builder::default()),
-        Box::new(impls::s3::Builder::default()),
-        Box::new(impls::s3_source::Builder::default()),
+        Box::new(impls::s3::s3::Builder::default()),
+        Box::new(impls::s3::s3_source::Builder::default()),
         Box::new(impls::kafka::consumer::Builder::default()),
     ]
 }
