+++ src/connectors/impls.rs
@@ -37,10 +37,7 @@ pub(crate) mod kv;
 pub(crate) mod metrics;
 /// Metronome
 pub(crate) mod metronome;
-
-
-pub(crate) mod s3_source;
-/// Connector for streaming to AWS S3
+/// Connectors for AWS S3
 pub(crate) mod s3;
 /// std streams connector (stdout, stderr, stdin)
 pub(crate) mod stdio;
